update:apitester

+++ scripts/apitester/setup.sh
@@ -6,8 +6,16 @@ G='\033[0;32m'
 NC='\033[0m'
 COOL_FOLDER="$HOME/.$USER.scripts/apitester"
 mkdir -p "$COOL_FOLDER"
-cp -r . $COOL_FOLDER
-echo -e "${G}Copied apitester files to $COOL_FOLDER ${NC}"
+if [ -f "$COOL_FOLDER/profile_data.py" ]
+then
+  DATA=$(cat "$COOL_FOLDER/profile_data.py")
+  cp -r . $COOL_FOLDER
+  echo "$DATA" > "$COOL_FOLDER/profile_data.py"
+  echo -e "${G}Profiles found: Retaining current profile_data.py${NC}"
+else
+  cp -r . $COOL_FOLDER
+fi
+echo -e "${G}Updated apitester files to $COOL_FOLDER ${NC}"
 if grep -q "apitester" ~/.bashrc
 then
   echo -e "Already found alias ${C}apitester${NC}"
